structural changes to logging and alignment
