+++ assemble_ott.py
@@ -1330,4 +1330,9 @@ names_of_interest = ['Ciliophora',
                      'Euxinia',
                      'Xiphonectes',
                      'Cylindrocarpon',
+                     'Macrophoma',
+                     'Tricellulortus peponiformis',
+                     'Dischloridium',
+                     'Gloeosporium',
+                     'Exaiptasia pallida',
                      ]
